+++ frontend/components/ui/feed/comment/Comments.module.scss
@@ -1,7 +1,3 @@
 .comment_container {
 	@apply border-r border-l border-b border-solid border-customDark px-4 py-8 flex flex-col gap-6;
-
-	.form {
-		@apply flex-1;
-	}
 }
